+++ src/main.rs
@@ -65,6 +65,7 @@ extern crate toml;
 pub mod types;
 
 mod logger;
+mod config;
 mod descriptor;
 mod environment;
 mod command;